test(db): split rls-policies.integration.test.ts by RLS scope
Split 2,473-LOC monolith into 7 focused files by RLS scoping pattern
plus a shared helper. Test count: 87 → 96. All files ≤500 LOC.

Closes db-5bu5.
